% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c.spec_qc}
\alias{c.spec_qc}
\title{Variable list for performing cluster analysis.}
\format{
A data frame with the following variables:
\describe{
  \item{Variable}{Variable name}
  \item{Description}{short description of variable}
  \item{warnGAM}{warnGAM}
  \item{warnWRTDS}{warnWRTDS}
  \item{Type}{Type}
  \item{Where}{Where}
  \item{DescriptionDetailed}{DescriptionDetailed}
}
}
\usage{
c.spec_qc
}
\description{
Variable list for performing cluster analysis using
  predictions from GAM results. (Future updates for WRTDS files.)
}
\keyword{datasets}
