% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c.specQC}
\alias{c.specQC}
\title{Cluster analysis specification QC data}
\format{
A data frame with 6 variables:
\describe{
  \item{Variable}{Variable name}
  \item{Description}{short description of variable}
  \item{Level}{Level}
  \item{warnGAM}{warnGAM}
  \item{reqGAMS}{reqGAMS}
  \item{warnWRTDS}{warnWRTDS}
}
}
\usage{
c.specQC
}
\description{
List of variables needed to be defined by user for running
  cluster analysis
}
\keyword{datasets}
