% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpCentrDist.R
\name{calcGrpEuclideanDist}
\alias{calcGrpEuclideanDist}
\title{calculate euclidean distances between groups}
\usage{
calcGrpEuclideanDist(c.spec, data, grp1)
}
\arguments{
\item{c.spec}{list that stores specifications for cluster analysis}

\item{data}{table of data to cluster. Rows are the items to be clustered and
columns represent the different variables.}

\item{grp1}{cluster results table}
}
\value{
table of results
}
\description{
calculate euclidean distances between groups
}
\examples{
\dontrun{
# TBD

}

}
\seealso{
\code{\link{calcQuanClass}}
}
