% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterYear.R
\name{calcWaterYear}
\alias{calcWaterYear}
\title{Calculate water year from given year and month}
\usage{
calcWaterYear(month, year)
}
\arguments{
\item{month}{month must be from 1-12}

\item{year}{year}
}
\value{
vector of water year
}
\description{
Water years begin on October 1 and are accorded the year
  associated with the ending day of the water year. For example, October 1,
  2016 is in the 2017 water year.
}
\examples{
# TBD

}
\seealso{
\code{\link{readTextFile}}
}
