% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcYearMonth.R
\name{calcYearMonth}
\alias{calcYearMonth}
\title{Add year and month to data table}
\usage{
calcYearMonth(data, dateCol = date, yearCol = year, monthCol = month)
}
\arguments{
\item{data}{Data table to analyze. Must have one column: \code{dateCol} which
contains a date to analyze}

\item{dateCol}{Column name that contains date, e.g., \code{dateCol =
SampleDate}}

\item{yearCol}{Column name to store year data, e.g., \code{yearCol =
SampleYear}}

\item{monthCol}{Column name to store month data, e.g., \code{monthCol =
SampleMonth}}
}
\value{
data table with values for year and month added
}
\description{
Based on a data table with a date field, add year and month 
  fields to the table.
}
\examples{
\dontrun{
dat0 <- tidyr::tibble( a = seq(as.Date('2022-06-10'), as.Date('2023-03-01'), by = "30 days"))

dat0 <- calcYearMonth(dat0
  , dateCol = a
  , yearCol = Year
  , monthCol = theMonth)

dat0$WY <- calcWaterYear(dat0$Year, dat0$theMonth)
}

}
\seealso{
\code{\link{readTextFile}} \code{\link{calcWaterYear}}
}
