% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbp_stat_list}
\alias{cbp_stat_list}
\title{Station vectors for selections portions of Chesapeake Bay
tidal waters.}
\format{
A list with the following variables:
\describe{
  \item{summary}{Summary table of abbreviations, descriptions and stations}            
  \item{cbp}{Main Bay Stations}            
  \item{u_cbp}{Upper Bay Stations}           
  \item{md_cbp}{MD Bay Stations}              
  \item{va_cbp}{VA Bay Stations}              
  \item{md_wt}{MD Western Tributary Stations}
  \item{md_et}{MD Eastern Tributary Stations}
  \item{va_we}{VA Western Embayment Stations}
  \item{va_ee}{VA Eastern Embayment Stations}
  \item{pax}{Patuxent Stations}            
  \item{pot}{Potomac Stations}             
  \item{rpp}{Rappahannock Stations}        
  \item{yrk}{York Stations}                
  \item{jms}{James Stations}               
  \item{ebe}{Elizabeth River Stations} 
}
}
\usage{
cbp_stat_list
}
\description{
Station vectors for selections portions of Chesapeake Bay
tidal waters.
}
\keyword{datasets}
