% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerData.R
\name{centerData}
\alias{centerData}
\title{Center data}
\usage{
centerData(
  data,
  id_col,
  prof_col,
  value_col,
  center_type = NA,
  center_col = NA
)
}
\arguments{
\item{data}{Data table to analyze. Must have two columns: \code{value_col}
which contains the values to transform and \code{id_col} which has the 
id variable.}

\item{id_col}{Column name that contains id variable (e.g., "by" variable)}

\item{prof_col}{Column name that contains prof variable}

\item{value_col}{Column name that contains values for transforming}

\item{center_type}{type of transformation to perform. If \code{center_type =
"meanadjust"}, then the values are adjusted by subtracting the mean. If
\code{transform_type = NA}, then an unchanged data table is returned.}

\item{center_col}{Column name to store transformed data. If left as NA, then
\code{value_col} will be overwritten.}
}
\value{
data table with transformed column
}
\description{
Center data
}
\details{
This function centers data from a table. Allowable transformations are
\itemize{
\item "meanadjust": subtract mean
\item "standnorm": z-score (subtract mean and divide by standard deviation)
\item "percent": percent (express response as a percent of the mean) 
}

Negative values in \code{value_col} are not permitted for "percent" option.
}
\examples{
\dontrun{
# TBD

}

}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
