% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClusterMaps.R
\name{clusterMap}
\alias{clusterMap}
\title{Wrapper function for mapPoints() function}
\usage{
clusterMap(
  grp_data,
  grp_lab,
  grp_col,
  leg_title = "Station Grouping",
  col_lat = "latitude",
  col_lon = "longitude",
  col_grp = "grpColor",
  leg_pos = "left",
  leg_labs = grp_lab,
  leg_cols = grp_col,
  file_layer = NA,
  zoom_buffer = NA,
  map_coord_ratio = 1.3,
  boo_tidal = TRUE
)
}
\arguments{
\item{grp_data}{cluster analysis results table}

\item{grp_lab}{group labels}

\item{grp_col}{group color}

\item{leg_title}{legend title}

\item{col_lat}{column name with latitude data}

\item{col_lon}{column name with longitude data}

\item{col_grp}{column with group}

\item{leg_pos}{legend position}

\item{leg_labs}{legend labels}

\item{leg_cols}{legend colors}

\item{file_layer}{file with base layer}

\item{zoom_buffer}{zoom buffer}

\item{map_coord_ratio}{map coordinate ratio}

\item{boo_tidal}{tidal indicator}
}
\value{
map object
}
\description{
Wrapper function for mapPoints() function
}
\details{
...
}
\examples{
\dontrun{
m <- clusterMap(...)
m2 <-clusterMapLabel(cmap = m
  , lat = 39.5, lon = -77.00
  , lab_txt = c("Station Grouping", "by", "TN Level")
  , x_space = 0, y_space = 0.1, lab_col <- "black", lab_size = 8)

}


}
\seealso{
\code{\link{clusterMap}}
}
