% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClusterMaps.R
\name{clusterMapLabel}
\alias{clusterMapLabel}
\title{Add a text annotation to map from clusterMap}
\usage{
clusterMapLabel(
  cmap,
  lat,
  lon,
  lab_txt,
  x_space = 0,
  y_space = 0,
  lab_col = "black",
  lab_size = 8,
  h_just = 0
)
}
\arguments{
\item{cmap}{map object from clusterMap}

\item{lat}{vertical location of annotation expressed as latitude}

\item{lon}{horizontal location of annotation expressed as longitude}

\item{lab_txt}{text of label. Can be a vector}

\item{x_space}{change in horizontal location between lines}

\item{y_space}{change in vertical location between lines}

\item{lab_col}{label color}

\item{lab_size}{label size}
}
\value{
map object
}
\description{
Add a text annotation to map from clusterMap
}
\details{
...
}
\examples{
\dontrun{
m <- clusterMap(...)
m2 <-clusterMapLabel(cmap = m
  , lat = 39.5, lon = -77.00
  , lab_txt = c("Station Grouping", "by", "TN Level")
  , x_space = 0, y_space = 0.1, lab_col <- "black", lab_size = 8)

}


}
\seealso{
\code{\link{clusterMap}}
}
