% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBasePred.R
\name{createBasePred}
\alias{createBasePred}
\title{Create base prediction data set for GAM clusters}
\usage{
createBasePred(
  start_year = 1990,
  end_year = year(Sys.Date()),
  month_grid = 1:12,
  day_grid = 15,
  month_adj = NA
)
}
\arguments{
\item{start_year}{Begin year of analysis (scalar)}

\item{end_year}{End year of analysis (scalar)}

\item{month_grid}{vector of months to include in analysis}

\item{day_grid}{days of month to make predictions}

\item{month_adj}{month adjustment to accommodate water year analyses}
}
\value{
data table with base prediction data set
\itemize{
\item year - year (calendar basis)   
\item month - month   
\item day - day of month   
\item date - date   
\item dyear - date expressed as decimal year   
\item doy - day of year (calendar basis)   
\item year_adj - year (adjusted based on month_adj)   
\item doy_adj - day of year (adjusted based on year_adj)  
}
}
\description{
Constructs the base prediction data set for GAM clusters and
  includes dates, days of year, decimal year, etc.
}
\details{
Set \code{month_grid = 1:12} to include all months of the year. \code{month_grid
= 6:9} would only include June through September.  

Set \code{day_grid = 15} to set up data set for the 15th of every month.
\code{day_grid = c(10,20)} would set up data set for the 10 and 20th of every month.

Set \code{month_adj = NA} results in analysis performed on a calendar year.
Setting \code{month_adj = c(10, 11, 12)} would result in analysis being set
up for a water year basis, making October 1 the first day of the year. This
is accomplished by computing year_adj as the calendar year plus 1 ("+1") for
months 10-12. If \code{month_adj = c(-1, -2, -3)} (note the negative values),
then year_adj is set to the calendar year minus 1 ("-1") for months 1-3. This
has the effect of making April 1 the first day of the year in this example.
}
\examples{
\dontrun{
basePred <- createBasePred(start_year = 2015
 , end_year = 2016
 , month_grid = 1:12
 , day_grid = c(10,20)
 , month_adj = c(10,11,12))
}

}
\seealso{
\code{\link{readTextFile}}
}
