% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPredGAM.R
\name{createPredGAM}
\alias{createPredGAM}
\title{Create prediction data}
\usage{
createPredGAM(c.spec)
}
\arguments{
\item{c.spec}{list of cluster analysis specifications}
}
\value{
data table with full prediction data set
\itemize{
\item station - 
\item wqParm - 
\item wqLayer - 
\item yearCal - year (calendar basis)   
\item year - year for cluster 
\item month - month   
\item day - day of month   
\item pred - prediction
}
}
\description{
Create prediction data
}
\details{
...
}
\examples{
\dontrun{
# TBD
}

}
\seealso{
\code{\link{readTextFile}}
}
