% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterLeaves.R
\name{deterLeaves}
\alias{deterLeaves}
\title{Summarize cluster analysis results into table}
\usage{
deterLeaves(dend, id_row_vec, grp_cnt)
}
\arguments{
\item{dend}{dendrogram object from dendextend}

\item{id_row_vec}{id_row elements (typically representing the rows of data
used in cluster analysis)}

\item{grp_cnt}{number of clusters}
}
\description{
Summarize cluster analysis results into table
}
\details{
...
}
\examples{
\dontrun{
# TBD

}

}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
