% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{extract}
\alias{extract}
\title{Extract components from list or table to parent environment}
\usage{
extract(listvar, v)
}
\arguments{
\item{listvar}{name of target list or table}

\item{v}{vector of variables to extract from list or table}
}
\description{
Extract components from list or table to parent environment
}
\details{
...
}
\examples{
\dontrun{
# TBD

}

}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
