% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{firstCap}
\alias{firstCap}
\title{capitalize first letter of character string}
\usage{
firstCap(cs)
}
\arguments{
\item{cs}{character string}
}
\value{
character
}
\description{
capitalize first letter of character string
}
\details{
...
}
\examples{
\dontrun{
# TBD

}

}
\seealso{
\code{\link{simpleCap}} \code{\link{vec.strg}} \code{\link{firstCapAll}}
}
\keyword{internal}
