% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{firstCapAll}
\alias{firstCapAll}
\title{capitalize first letter of words separated by blanks in a character string}
\usage{
firstCapAll(cs)
}
\arguments{
\item{cs}{character string}
}
\value{
character
}
\description{
capitalize first letter of words separated by blanks in a
  character string and checks to make sure other letters are lower case
}
\details{
...
}
\examples{
\dontrun{
# TBD

}

}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
