% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{getColors}
\alias{getColors}
\title{Return vector of named colors}
\usage{
getColors(grp_cnt)
}
\arguments{
\item{k}{number of colors}
}
\value{
named vector
}
\description{
Return vector of named colors
}
\examples{

print(hexColor2Name("#80FF00"))

grp_cnt <- 5

grp_col <- rev(rainbow(grp_cnt)) 
grp_col <- apply(as.data.frame(grp_col), 1, hexColor2Name)
barplot(1:grp_cnt, col=grp_col, names.arg=grp_col)

# baycluster colors
grp_col = scales::hue_pal()(grp_cnt)
grp_col = apply(as.data.frame(grp_col), 1, hexColor2Name)
barplot(1:grp_cnt, col=grp_col, names.arg=grp_col)

# simpler baycluster color call statement
grp_cnt <- 4
grp_col <- getColors(grp_cnt)
barplot(1:grp_cnt, col=grp_col, names.arg=grp_col)

}
\seealso{
\code{\link{hexColor2Name}} \code{\link[scales]{hue_pal}}
}
\keyword{internal}
