% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{getColors2}
\alias{getColors2}
\title{Return vector of named colors ... red to blue}
\usage{
getColors2(k)
}
\arguments{
\item{k}{number of colors}
}
\value{
named vector
}
\description{
Return vector of named colors  ... red to blue
}
\examples{

k <- 5

grp_col <- rev(rainbow(k))
grp_col <- apply(as.data.frame(grp_col), 1, hexColor2Name)
barplot(1:k, col=grp_col, names.arg=grp_col)

# baycluster color call statement
k <- 5
grp_col <- getColors(k)
barplot(1:k, col=grp_col, names.arg=grp_col)

# baycluster color call statement for exogenous covariate
k <- 4
grp_col <- getColors2(k)
barplot(1:k, col=grp_col, names.arg=grp_col)

}
\seealso{
\code{\link{hexColor2Name}} \code{\link{getColors}} \code{\link[scales]{hue_pal}}
}
\keyword{internal}
