% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{grabFunctionArguments}
\alias{grabFunctionArguments}
\title{Create a list of all arguments passed to function}
\usage{
grabFunctionArguments()
}
\value{
list of arguments
}
\description{
Create a list of all arguments passed to function
}
\details{
Derived from https://stackoverflow.com/questions/66329835/using-r-how-to-get-all-parameters-passed-into-a-function-with-their-values
}
\examples{
\dontrun{

my_function = function(z, pi_0 = 0.3, families=list(), ... ) {
  args = grabFunctionArguments()
  names(args)
  return(args)
}

X <- my_function(z=4, a=345, families = list(a=34, b=545), myList = list(a=34, b=545))
}

}
\keyword{internal}
