% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{hexColor2Name}
\alias{hexColor2Name}
\title{Return named color close to hexCol}
\usage{
hexColor2Name(hexCol)
}
\arguments{
\item{hexCol}{hex color formatted character, i.e., "#80FF00"}
}
\value{
named color
}
\description{
Return named color close to hexCol
}
\examples{

print(hexColor2Name("#80FF00"))

grp_cnt <- 5

grp_col <- rev(rainbow(grp_cnt)) 
grp_col <- apply(as.data.frame(grp_col), 1, hexColor2Name)
barplot(1:grp_cnt, col=grp_col, names.arg=grp_col)

# baycluster colors
grp_col = scales::hue_pal()(grp_cnt)
grp_col = apply(as.data.frame(grp_col), 1, hexColor2Name)
barplot(1:grp_cnt, col=grp_col, names.arg=grp_col)

# simpler baycluster color call statement
grp_cnt <- 4
grp_col <- getColors(grp_cnt)
barplot(1:grp_cnt, col=grp_col, names.arg=grp_col)


}
\seealso{
\code{\link{getColors}} \code{\link[scales]{hue_pal}}
}
\keyword{internal}
