% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterMaps.R
\name{mapPoints}
\alias{mapPoints}
\title{Basic map plotting function}
\usage{
mapPoints(
  pts,
  col_lat = "latitude",
  col_lon = "longitude",
  col_grp = NA,
  leg_pos = "bottom",
  leg_title = NULL,
  leg_labs = NULL,
  leg_cols = NULL,
  file_layer = file_layer,
  zoom_buffer = NA,
  map_coord_ratio = 1.3,
  boo_tidal = TRUE
)
}
\arguments{
\item{pts}{table of data to plot on a map}

\item{col_lat}{column name with latitude data}

\item{col_lon}{column name with longitude data}

\item{col_grp}{column name with group colors}

\item{leg_pos}{legend position}

\item{leg_title}{legend title}

\item{leg_labs}{legend labels}

\item{leg_cols}{legend colors}

\item{file_layer}{directory with base layer  (note: I’ll probably switch to the file.path at a later time)}

\item{zoom_buffer}{zoom layer}

\item{map_coord_ratio}{map coordinate ratio}

\item{boo_tidal}{flag to indicate tidal}
}
\value{
map object
}
\description{
Basic map plotting function
}
\details{
...
}
\examples{
\dontrun{
m <- plotMap(...)
m2 <-plotMapLabel(cmap = m
  , lat = 39.5, lon = -77.00
  , lab_txt = c("Station Grouping", "by", "TN Level")
  , x_space = 0, y_space = 0.1, lab_col <- "black", lab_size = 8)

}

}
\seealso{
\code{\link{plotMap}}
}
\keyword{internal}
