% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterMaps.R
\name{plotMap}
\alias{plotMap}
\title{Wrapper function for mapPoints() function}
\usage{
plotMap(
  grp_data,
  col_lbl = "prim_lbl",
  col_colors = "prim_col",
  leg_title = "Groups",
  leg_pos = "left",
  col_lat = "latitude",
  col_lon = "longitude",
  file_layer = NA,
  zoom_buffer = NA,
  map_coord_ratio = 1.3,
  boo_tidal = TRUE
)
}
\arguments{
\item{grp_data}{table of data to be plotted}

\item{col_lbl}{column with labels}

\item{col_colors}{column with colors}

\item{leg_title}{legend title}

\item{leg_pos}{legend position}

\item{col_lat}{column name with latitude data}

\item{col_lon}{column name with longitude data}

\item{file_layer}{file with base layer}

\item{zoom_buffer}{zoom buffer}

\item{map_coord_ratio}{map coordinate ratio}

\item{boo_tidal}{tidal indicator}
}
\value{
map object
}
\description{
Wrapper function for mapPoints() function
}
\details{
...
}
\examples{
\dontrun{
m <- plotMap(...)
m2 <-plotMapLabel(cmap = m
  , lat = 39.5, lon = -77.00
  , lab_txt = c("Station Grouping", "by", "TN Level")
  , x_space = 0, y_space = 0.1, lab_col <- "black", lab_size = 8)

}


}
\seealso{
\code{\link{plotMap}}
}
