% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfiles.R
\name{pltClusProfMat}
\alias{pltClusProfMat}
\title{Plot profiles - one plot}
\usage{
pltClusProfMat(c.spec, data, grp1, legendPos = "topleft")
}
\arguments{
\item{c.spec}{list that stores specifications for cluster analysis}

\item{data}{table of data to plot. Rows are the items to plot and
columns represent the profile data.}

\item{grp1}{cluster results table}

\item{legendPos}{legend position}
}
\value{
table of plotting data
}
\description{
Plot profiles - one plot
}
\examples{
\dontrun{
#TBD

}

}
\seealso{
\code{\link{calcQuanClass}}
}
