% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfiles.R
\name{pltClusProfMatPan}
\alias{pltClusProfMatPan}
\title{Plot profiles - panels}
\usage{
pltClusProfMatPan(c.spec, data, grp1, legendPos = "topleft", num_col = 3)
}
\arguments{
\item{c.spec}{list that stores specifications for cluster analysis}

\item{data}{table of data to plot. Rows are the items to plot and
columns represent the profile data.}

\item{grp1}{cluster results table}

\item{legendPos}{legend position}

\item{num_col}{number of columns}
}
\value{
table of plotting data
}
\description{
Plot profiles - panels
}
\examples{
\dontrun{
#TBD

}

}
\seealso{
\code{\link{calcQuanClass}}
}
