% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{pry}
\alias{pry}
\title{pry components from list or table to parent environment}
\usage{
pry(listvar, v = NA)
}
\arguments{
\item{listvar}{name of target list or table}

\item{v}{vector of variables to pry from list or table}
}
\description{
pry components from list or table to parent environment
}
\details{
...
}
\examples{

my_list <- list(a = 4, b = 5, c = "charlie")
pry(my_list, c("a", "c"))

d = "delta"
my_list <- stow(my_list, v = "d")
my_list

}
\seealso{
\code{\link{stow}}
}
\keyword{internal}
