% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTextFile.R
\name{readTextFile}
\alias{readTextFile}
\title{Read text file into variable}
\usage{
readTextFile(
  fileName,
  filePath = "..",
  fileType = "standard",
  colNames = NA,
  report = TRUE
)
}
\arguments{
\item{fileName}{Name of input file as character array}

\item{filePath}{File path as character array. Setting \code{filePath = ".."}
reads from the current working directory}

\item{fileType}{Indicates what type of file structure is expected. Set
\code{fileType = "WRTDS"} or \code{fileType = "flow"} to use specialized
settings built into the function.  \code{fileType = "standard"} is the default
setting for typical comma or tab delimited text files with the header
as the first line.}

\item{colNames}{Vector of columns names to use for column names in the
returned data set. Set \code{colNames = NA} to use column headers from
input file.}

\item{report}{Summary information about data after reading. Set
\code{report = FALSE} to suppress information.}
}
\value{
data table
}
\description{
This function facilitates reading in text files with standard 
and non-standard headers. The two non-standard file types are 1) USGS' WRTDS
reporting file for estimated concentrations and loads and 2) USGS
flow retrievals.
}
\details{
...
}
\examples{
# TBD

}
\seealso{
\code{\link{calcQuanClass}}
}
