% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupLabeling.R
\name{rnkFlowCorLab}
\alias{rnkFlowCorLab}
\title{compute group labels based on correlation with flow}
\usage{
rnkFlowCorLab(c.spec, allMat.mn)
}
\arguments{
\item{c.spec}{list that stores specifications for cluster analysis}

\item{allMat.mn}{table of profile means}
}
\value{
table of results
}
\description{
compute group labels based on correlation with flow
}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
