% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupLabeling.R
\name{rnkTrndLab}
\alias{rnkTrndLab}
\title{compute group labels based on comparison of beginning and ending 3-yr time period}
\usage{
rnkTrndLab(c.spec, allMat.mn)
}
\arguments{
\item{c.spec}{list that stores specifications for cluster analysis}

\item{allMat.mn}{table of profile means}
}
\value{
table of results
}
\description{
compute group labels based on comparison of beginning and ending 3-yr time period
}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
