% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSpec.R
\name{setSpec}
\alias{setSpec}
\title{Create (or update) specification list for cluster analysis}
\usage{
setSpec(c.spec = list(), ...)
}
\arguments{
\item{c.spec}{list for storing specifications for cluster analysis}

\item{...}{specifications needed to run cluster analysis -- see details for
list of needed variables}
}
\value{
list
}
\description{
Create (or update) specification list for cluster analysis.
  Currently developed for using predictions from GAM results. (Future updates
  for WRTDS files.)
}
\details{
See \code{\link{c.spec_qc}} for list of variables needed to be passed. 

TBD: expand to include list of variables added to the list
}
\examples{
\dontrun{
c.spec <- setSpec(c.spec = list()
  , start_year    = 1993
  , end_year      = 2020
  , month_grid    = 1:12
  , day_grid      = 15 
  , month_adj     = NA
  ) 

}

}
\seealso{
\code{\link{calcQuanClass}}
}
