% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSpecBuildout.R
\name{setSpecBuildout}
\alias{setSpecBuildout}
\title{Add supporting variables to c.spec}
\usage{
setSpecBuildout(c.spec)
}
\arguments{
\item{c.spec}{list of cluster analysis specifications}
}
\value{
n/a
}
\description{
Add supporting variables to c.spec
}
\details{
...
}
\examples{
# TBD

}
\seealso{
\code{\link{readTextFile}}
}
\keyword{internal}
