% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSpecCheck.R
\name{setSpecCheck}
\alias{setSpecCheck}
\title{Perform rudimentary data checking of c.spec}
\usage{
setSpecCheck(c.spec)
}
\arguments{
\item{c.spec}{list of cluster analysis specifications}
}
\value{
n/a
}
\description{
Perform rudimentary data checking of c.spec
}
\details{
...
}
\examples{
# TBD

}
\seealso{
\code{\link{readTextFile}}
}
\keyword{internal}
