% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSpecChk.R
\name{setSpecChk}
\alias{setSpecChk}
\title{Perform rudimentary data checking of variables in c.spec}
\usage{
setSpecChk(c.spec)
}
\arguments{
\item{c.spec}{list of cluster analysis specifications}
}
\description{
Perform rudimentary data checking of variables in c.spec. Report
  out which variables should have been passed but not present.
}
\details{
...
}
\examples{
# TBD


}
\seealso{
\code{\link{readTextFile}}
}
\keyword{internal}
