% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSpecChkGAM.R
\name{setSpecChkGAM}
\alias{setSpecChkGAM}
\title{Perform GAM-specific data checking of c.spec}
\usage{
setSpecChkGAM(c.spec)
}
\arguments{
\item{c.spec}{list of cluster analysis specifications}
}
\description{
Perform GAM-specific data checking of c.spec
}
\details{
...
}
\examples{
# TBD


}
\seealso{
\code{\link{readTextFile}}
}
\keyword{internal}
