% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSpecChkWRTDS.R
\name{setSpecChkWRTDS}
\alias{setSpecChkWRTDS}
\title{Perform WRTDS-specific data checking of c.spec}
\usage{
setSpecChkWRTDS(c.spec)
}
\arguments{
\item{c.spec}{list of cluster analysis specifications}
}
\description{
Perform WRTDS-specific data checking of c.spec
}
\details{
...
}
\examples{
# TBD


}
\seealso{
\code{\link{readTextFile}}
}
\keyword{internal}
