% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSpecCmp.R
\name{setSpecCmp}
\alias{setSpecCmp}
\title{Complete c.spec with supporting variables}
\usage{
setSpecCmp(c.spec)
}
\arguments{
\item{c.spec}{list of cluster analysis specifications}
}
\value{
list
}
\description{
Complete c.spec with supporting variables
}
\details{
...
}
\examples{
# TBD

}
\seealso{
\code{\link{readTextFile}}
}
\keyword{internal}
