% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSpecCmpGAM.R
\name{setSpecCmpGAM}
\alias{setSpecCmpGAM}
\title{Complete c.spec with GAM-specific supporting variables}
\usage{
setSpecCmpGAM(c.spec)
}
\arguments{
\item{c.spec}{list of cluster analysis specifications}
}
\value{
list
}
\description{
Complete c.spec with GAM-specific supporting variables
}
\details{
...
}
\examples{
# TBD

}
\seealso{
\code{\link{readTextFile}}
}
\keyword{internal}
