% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSpecCmpWRTDS.R
\name{setSpecCmpWRTDS}
\alias{setSpecCmpWRTDS}
\title{Complete c.spec with WRTDS-specific supporting variables}
\usage{
setSpecCmpWRTDS(c.spec)
}
\arguments{
\item{c.spec}{list of cluster analysis specifications}
}
\value{
list
}
\description{
Complete c.spec with WRTDS-specific supporting variables
}
\details{
...
}
\examples{
# TBD

}
\seealso{
\code{\link{readTextFile}}
}
\keyword{internal}
