% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_cbseg}
\alias{shp_cbseg}
\title{Base layer for making maps in the tidal waters of the 
Chesapeake Bay.}
\format{
A data frame with the following variables:
\describe{
  \item{long}{longitude}
  \item{lat}{latitude}
  \item{order}{order with in a piece}
  \item{hole}{hole}
  \item{piece}{piece}
  \item{id}{id}
  \item{group}{group}
}
}
\usage{
shp_cbseg
}
\description{
Base layer for making maps in the tidal waters of the 
Chesapeake Bay.
}
\keyword{datasets}
