% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{simpleCap}
\alias{simpleCap}
\title{capitalize first letter of words separated by blanks in a character string}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{character string}
}
\value{
character
}
\description{
capitalize first letter of words separated by blanks in a character string
}
\details{
...
}
\examples{
\dontrun{
# TBD

}

}
\seealso{
\code{\link{vec.strg}} \code{\link{firstCap}} \code{\link{firstCapAll}}
}
\keyword{internal}
