% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{stow}
\alias{stow}
\title{append variables to a list}
\usage{
stow(listvar = list(), v = NA)
}
\arguments{
\item{listvar}{name of target list}

\item{v}{vector of variables to stow into list}
}
\description{
append variables to a list
}
\details{
...
}
\examples{
\dontrun{

a = c(1, 3, 5)
b = c("abc", "def", "ghi", "jkl")
my_list <- stow(my_list, c("a", "b") )

}

}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
