% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblFT1.R
\name{tblFT1}
\alias{tblFT1}
\title{Output formatted table.}
\usage{
tblFT1(
  data,
  tbl_title = NA,
  tbl_pre_label = NA,
  tbl_font_name = NA,
  tbl_font_size = NA,
  tbl_theme = NA
)
}
\arguments{
\item{data}{data set to output}

\item{tbl_title}{table title}

\item{tbl_pre_label}{table title pre_label, e.g., "Table: "}

\item{tbl_font_name}{table font, e.g., "Calibri"}

\item{tbl_font_size}{font size}

\item{tbl_theme}{table theme, available options: "box", "vanilla", "booktabs"}
}
\value{
# may return flextable
}
\description{
This function outputs a table of data with basic formatting.
}
\details{
...
}
\examples{

tblFT1(head(iris))

tblFT1(head(iris)
  , tbl_theme ="box")
  

}
\seealso{
\code{\link{calcQuanClass}}
}
