% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformData.R
\name{transformData}
\alias{transformData}
\title{Transform column of data}
\usage{
transformData(
  data,
  value_col,
  transform_type = "logtransform",
  transform_col = NA
)
}
\arguments{
\item{data}{Data table to analyze. Must have one column: \code{value_col}
which contain the values to transform}

\item{value_col}{Column name that contains values for transforming}

\item{transform_type}{type of transformation to perform. If \code{transform_type =
"logtransform"}, then values are log transformed}

\item{transform_col}{Column name to store transformed data. If left as NA, then
value_col will be overwritten.}
}
\value{
data table with transformed column
}
\description{
Transform column of data
}
\details{
This function transforms a column of data from a table. Allowable transformations are
\itemize{
\item "logtransform": natural log transformation
\item "exptransform": exp(x) transform
\item "logbase10": log base 10 
\item "antilog10": 10^ (anti-log for base 10 logarithms)
\item "standnorm": z-score (subtract mean and divide by standard deviation)
\item "percent": percent (express response as a percent of the mean)
}
}
\examples{
\dontrun{
# TBD

}

}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
