% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{vec.strg}
\alias{vec.strg}
\title{Convert vector to a single character}
\usage{
vec.strg(x, sep = ", ")
}
\arguments{
\item{x}{vector to be strung together}

\item{sep}{separator between elements of vector}
}
\value{
character
}
\description{
Convert vector to a single character
}
\details{
...
}
\examples{
\dontrun{
M <- 1:10
vec.strg(M)

}

}
\seealso{
\code{\link{simpleCap}} \code{\link{firstCap}} \code{\link{firstCapAll}}
}
\keyword{internal}
