% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupLabeling.R
\name{autogrp_lab}
\alias{autogrp_lab}
\title{compute mean group profiles and assign group labels}
\usage{
autogrp_lab(c.spec, allMat, grp1)
}
\arguments{
\item{c.spec}{list that stores specifications for cluster analysis}

\item{allMat}{table of data to cluster. Rows are the items to be clustered and
columns represent the different variables.}

\item{grp1}{cluster results table}
}
\value{
table of results
}
\description{
compute mean group profiles and assign group labels
}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
