% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQuanClass.R
\name{calcQuanClass}
\alias{calcQuanClass}
\title{Calculate and return quantile class by year and month}
\usage{
calcQuanClass(
  data,
  dateCol = "date",
  valueCol = "flow",
  transform = TRUE,
  probabilities = seq(0, 1, 0.25),
  startYear,
  endYear,
  yearType = "calendar"
)
}
\arguments{
\item{data}{Data table to analyze. Must have two columns: \code{dateCol} and
\code{valueCol} which contain a date and value to analyze, respectively}

\item{dateCol}{Column name that contains date}

\item{valueCol}{Column name that contains values to analyze for quantiles}

\item{transform}{Log transform data before analysis}

\item{probabilities}{Vector of probabilities to analyze}

\item{startYear}{First year to maintain in returned data set}

\item{endYear}{Last year to maintain in returned data set}

\item{yearType}{"calendar" or "water" year}
}
\value{
data table with cross tabulation of quantiles by year and month
}
\description{
Calculate and return quantile class by year and month
}
\details{
Question: Should quantiles be computed with all data and then trim
un-used years data? Or should data be trimmed by year first?
}
\examples{
# TBD

}
\seealso{
\code{\link{readTextFile}}
}
