% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTotWSS.R
\name{calcTotWSS}
\alias{calcTotWSS}
\title{Calculate total within sum of squares}
\usage{
calcTotWSS(data, dist_method = "euclidean", aggl_method = "ward.D")
}
\arguments{
\item{data}{table of data to cluster. Rows are the items to be clustered and
columns represent the different variables.}

\item{dist_method}{distance matrix method to be used in coordination with
stats::dist(). This must be one of "euclidean", "maximum", "manhattan",
"canberra", "binary" or "minkowski".}

\item{aggl_method}{dissimilarities agglomeration method to be used in
coordination with stats::hclust(). This should be one of "ward.D",
"ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA),
"median" (= WPGMC) or "centroid" (= UPGMC)}
}
\value{
"knee of the curve" plot
}
\description{
Calculate total within sum of squares to inform the number of
  clusters to use. Create and return "knee-of-the-curve" plot
}
\details{
The "knee-of-the-curve" method can be used to inform the number of
  clusters to use when performing cluster analyses. The input data should
  only have columns used in the cluster analysis.
}
\examples{
\dontrun{
calTotWSS(iris[ , -5]) 
}

}
\seealso{
\code{\link{calcQuanClass}}
}
