% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterYear.R
\name{calcWaterYear}
\alias{calcWaterYear}
\title{Calculate water year from given year and month}
\usage{
calcWaterYear(yearIn, monthIn)
}
\arguments{
\item{yearIn}{year}

\item{monthIn}{month must be from 1-12}
}
\value{
vector of water year
}
\description{
Water years begin on October 1 and are accorded the year
  associated with the ending day of the water year. For example, October 1,
  2016 is in the 2017 water year.
}
\examples{
\dontrun{
dat0 <- tidyr::tibble( a = seq(as.Date('2022-06-10'), as.Date('2023-03-01'), by = "30 days"))

dat0 <- calcYearMonth(dat0
  , dateCol = "a"
  , yearCol = "Year"
  , monthCol = "theMonth")

dat0$WY <- calcWaterYear(dat0$Year, dat0$theMonth)
}

}
\seealso{
\code{\link{readTextFile}} \code{\link{calcYearMonth}}
}
