% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWaterYear.R
\name{calcWaterYear}
\alias{calcWaterYear}
\title{Add water year from given year and month to a data table}
\usage{
calcWaterYear(data, yearCol = "year", monthCol = "month", wateryearCol = NA)
}
\arguments{
\item{data}{Data table to analyze. Must have two column: \code{yearCol} and
\code{monthCol}}

\item{yearCol}{Column name that contains year data}

\item{monthCol}{Column name that contains month data}

\item{wateryearCol}{Column name to store water year. If left as NA, then
yearCol will be overwritten.}
}
\value{
vector of water year
}
\description{
Water years begin on October 1 and are accorded the year
  associated with the ending day of the water year. For example, October 1,
  2016 is in the 2017 water year.
}
\examples{
\dontrun{
dat0 <- tidyr::tibble( a = seq(as.Date('2022-06-10'), as.Date('2023-03-01'), by = "30 days"))

dat0 <- calcYearMonth(dat0
  , dateCol = "a"
  , yearCol = "Year"
  , monthCol = "theMonth")

dat0 <- calcWaterYear(dat0, "Year", "theMonth", "WY")
}

}
\seealso{
\code{\link{readTextFile}} \code{\link{calcYearMonth}}
}
\keyword{internal}
