% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkRDAfiles.R
\name{chkRDAfiles}
\alias{chkRDAfiles}
\title{Check RDA files to confirm availability of GAM results}
\usage{
chkRDAfiles(
  c.spec = NULL,
  station_vec = NULL,
  wq_parm = NULL,
  wq_layer = NULL,
  gam_numbr = NULL,
  start_year = NULL,
  end_year = NULL,
  gam_folder = NULL,
  month_threshold = NULL
)
}
\arguments{
\item{c.spec}{cluster specification file}

\item{station_vec}{Vector of stations to analyze}

\item{wq_parm}{Parameter abbreviation to evaluate}

\item{wq_layer}{Layer abbreviation to evaluate}

\item{gam_numbr}{GAM option to evaluate}

\item{start_year}{Begin year of analysis (scalar)}

\item{end_year}{End year of analysis (scalar)}

\item{gam_folder}{Folder location of GAM results from baytrends}

\item{month_threshold}{Threshold for flagging acceptable data range (see details)}
}
\value{
data table with a summary of files and GAM results
}
\description{
Based on a station list, year range, water quality variable,
  layer, and GAM option, confirm that the `gamResults` from
  baytrends are correctly positioned.
}
\details{
It is presumed that a series of baytrends operations
have been run and the results for an individual "station|parameter|layer"
have been stored as separate rda files in a folder passed as the variable
\code{gam_folder} as depicted below.

\code{gam_folder/tn/TF3.1E_tn_surf.rda}\cr  
\code{gam_folder/tn/TF3.2E_tn_surf.rda}\cr 
\code{gam_folder/tn/...}\cr
\code{gam_folder/do/TF3.1E_do_surf.rda}\cr
\code{gam_folder/do/TF3.2E_do_surf.rda}\cr 
\code{gam_folder/do/...}\cr
           
The values of \code{station}, \code{wq_parm}, and \code{wq_layer} are used to 
construct the file name convention of: "station_parameter_layer.rda". Under the 
folder \code{gam_folder}, subfolders are organized by \code{wq_parm}. 

\code{gam_numbr} is used to define which GAM formula to use in the analysis.
Setting \code{gam_numbr = c(23, 3, 2)} would result in preferentially
selecting \code{gamOutput23}, then \code{gamOutput3}, and finally \code{gamOutput2}.

\code{start_year} and \code{end_year} are the planned beginning and ending
period of the cluster analysis. This function checks to confirm that a 
particular combination of \code{station}, \code{wq_parm}, and \code{wq_layer} 
had monitoring data for the full period or record, i.e., \code{start_year}-01-01 
to \code{end_year}-12-31. 

\code{month_threshold} provides a grace period in the interpretation of a full
period of record.
}
\examples{
\dontrun{
# TBD

}

}
\seealso{
\code{\link{readTextFile}}
}
