% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterData.R
\name{clusterData}
\alias{clusterData}
\title{Perform basic hierarchical cluster analysis}
\usage{
clusterData(
  c.spec = NULL,
  data,
  man_dend_grp_lbl = NA,
  dend_title = NA,
  dend_xlab = NA,
  dend_ylab = NA,
  ...
)
}
\arguments{
\item{c.spec}{list that stores specifications for cluster analysis}

\item{data}{table of data to cluster. Rows are the items to be clustered and
columns represent the different variables. See Details for more information.}

\item{man_dend_grp_lbl}{Labels to manually override automatic group labeling}

\item{dend_title}{user supplied figure title}

\item{dend_xlab}{user supplied x axis label}

\item{dend_ylab}{user supplied y axis label}

\item{...}{alternative variable passing}
}
\value{
table of results
}
\description{
Perform basic hierarchical cluster analysis and return an
  object which describes the produced tree.
}
\details{
This function calculates the distance matrix between the rows of a
  data matrix. This distance matrix is then used to perform the hierarchical
  cluster analysis. 
  
The input \code{data} must be organized to have the first column as an
identifier for the rows. The remainder of the columns have data to be used in
the analysis. It is strongly advised and advantageous to have labeled rows
for plotting purposes. The output from \code{crossTabulate} is properly 
formatted as input.

The variables actively used from \code{c.spec} include the following:

\code{dist_method} distance matrix method to be used in coordination with
  stats::dist(). This must be one of "euclidean", "maximum", "manhattan",
  "canberra", "binary" or "minkowski".
  
\code{aggl_method} dissimilarities agglomeration method to be used in
  coordination with stats::hclust(). This should be one of "ward.D",
  "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA),
  "median" (= WPGMC) or "centroid" (= UPGMC)

\code{wq_parm} 
\code{grp_cnt}
\code{prof_var}
\code{id_var}
}
\examples{
\dontrun{
#TBD

}

}
\seealso{
\code{\link{crossTabulate}} \code{\link[stats]{dist}} \code{\link[stats]{hclust}}
}
