% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBasePred.R
\name{createBasePred}
\alias{createBasePred}
\title{Create base prediction data set for GAM clusters}
\usage{
createBasePred(
  startYear = 1990,
  endYear = year(Sys.Date()),
  monthGrid = 1:12,
  dayGrid = 15,
  month.adj = NA
)
}
\arguments{
\item{startYear}{Begin year of analysis (scalar)}

\item{endYear}{End year of analysis (scalar)}

\item{monthGrid}{vector of months to include in analysis}

\item{dayGrid}{days of month to make predictions}

\item{month.adj}{month adjustment to accommodate water year analyses}
}
\value{
data table with base prediction data set
\itemize{
\item year - year (calendar basis)   
\item month - month   
\item day - day of month   
\item date - date   
\item dyear - date expressed as decimal year   
\item doy - day of year (calendar basis)   
\item year.adj - year (adjusted based on month.adj)   
\item doy.adj - day of year (adjusted based on year.adj)  
}
}
\description{
Constructs the base prediction data set for GAM clusters and
  includes dates, days of year, decimal year, etc.
}
\details{
Set \code{monthGrid = 1:12} to include all months of the year. \code{monthGrid
= 6:9} would only include June through September.  

Set \code{dayGrid = 15} to set up data set for the 15th of every month.
\code{dayGrid = c(10,20)} would set up data set for the 10 and 20th of every month.

Set \code{month.adj = NA} results in analysis performed on a calendar year.
Setting \code{month.adj = c(10, 11, 12)} would result in analysis being set
up for a water year basis, making October 1 the first day of the year. This
is accomplished by computing year.adj as the calendar year plus 1 ("+1") for
months 10-12. If \code{month.adj = c(-1, -2, -3)} (note the negative values),
then year.adj is set to the calendar year minus 1 ("-1") for months 1-3. This
has the effect of making April 1 the first day of the year in this example.
}
\examples{
\dontrun{
basePred <- createBasePred(startYear = 2015
 , endYear = 2016
 , monthGrid = 1:12
 , dayGrid = c(10,20)
 , month.adj = c(10,11,12))
}

}
\seealso{
\code{\link{readTextFile}}
}
