% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossTabulate.R
\name{crossTabulate}
\alias{crossTabulate}
\title{Cross tabulate data.}
\usage{
crossTabulate(c.spec, data, ret_data = 1)
}
\arguments{
\item{c.spec}{list for storing specifications for cluster analysis}

\item{data}{input data to be cross tabulated}

\item{ret_data}{1: return wide data, 2: return list of wide and long data}
}
\value{
cross tabulate data table
}
\description{
Cross tabulate data. Includes transformations, averaging over
  id_var., and data centering
}
\details{
The incoming data are intended to have the following fields:
  "station", "wq_parm",  "wq_layer", "yearCal", "year", "month",  "day"
  "value". JBH NOTE: need to clarify minimum data set and what happens with
  extra variables.

Processing include the following steps:  

Values are transformed based on value stored in c.spec$data_transform

If c.spec$id_var identifies more than one variable, a new column is
constructed by concatenating the columns identified by id_var.

Values are averaged over id_var (or the newly constructed column).

Values are centered based on the value of c.spec$data_center and performed
on an id_var basis

Values are cross tabulated such that id_var is unique and columns correspond
to prof_var
}
\examples{
\dontrun{
#TBD

}

}
\seealso{
\code{\link{calcQuanClass}} \code{\link{transformData}} \code{\link{centerData}}
}
