% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{getColors}
\alias{getColors}
\title{Return vector of named colors}
\usage{
getColors(grp_cnt)
}
\arguments{
\item{k}{number of colors}
}
\value{
named vector
}
\description{
Return vector of named colors
}
\details{
...
}
\examples{
\dontrun{

hexColor2Name("#80FF00")

grp_cnt <- 5

grp_col <- rev(rainbow(grp_cnt))

grp_df  <- tibble(lab = paste("Group",1:grp_cnt)) \%>\%
  mutate(., grp_col = scales::hue_pal()(grp_cnt)) \%>\%
  mutate(., grp_col = apply(as.data.frame(grp_col), 1, hexColor2Name)) 
with(grp_df, barplot(1:grp_cnt, col=grp_col, names.arg=grp_col))

getColors(5)

}

}
\seealso{
\code{\link{hexColor2Name}}
}
\keyword{internal}
