% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{hexColor2Name}
\alias{hexColor2Name}
\title{Find named color close to hexCol}
\usage{
hexColor2Name(hexCol)
}
\arguments{
\item{hexCol}{hex color formatted character, i.e., "#80FF00"}
}
\value{
named color
}
\description{
Find named color close to hexCol
}
\details{
...
}
\examples{
\dontrun{

hexColor2Name("#80FF00")

grp_cnt <- 5

grp_col <- rev(rainbow(grp_cnt))

grp_df  <- tibble(lab = paste("Group",1:grp_cnt)) \%>\%
  mutate(., grp_col = scales::hue_pal()(grp_cnt)) \%>\%
  mutate(., grp_col = apply(as.data.frame(grp_col), 1, hexColor2Name)) 
with(grp_df, barplot(1:grp_cnt, col=grp_col, names.arg=grp_col))
}

}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
