% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{hexColor2Name}
\alias{hexColor2Name}
\title{Find named color close to hexCol}
\usage{
hexColor2Name(hexCol)
}
\arguments{
\item{hexCol}{hex color formatted character, i.e., "#80FF00"}
}
\value{
named color
}
\description{
Find named color close to hexCol
}
\details{
...
}
\examples{
\dontrun{

hexColor2Name("#80FF00")

grpCnt <- 5

grpCol <- rev(rainbow(grpCnt))

grpDF  <- tibble(lab = paste("Group",1:grpCnt)) \%>\%
  mutate(., grpCol = scales::hue_pal()(grpCnt)) \%>\%
  mutate(., grpCol = apply(as.data.frame(grpCol), 1, hexColor2Name)) 
with(grpDF, barplot(1:grpCnt, col=grpCol, names.arg=grpCol))
}

}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
