% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outSummStat.R
\name{outSummStat}
\alias{outSummStat}
\title{Output summary statistics about table into table}
\usage{
outSummStat(data, tblFTout = FALSE)
}
\arguments{
\item{data}{table variable}

\item{tblFTout}{set to TRUE to output formatted tables}
}
\value{
# returns flextable list
}
\description{
This function outputs the 5-number summary for numeric and
date fields. Character arrays are summarized.
}
\details{
...
}
\examples{
\dontrun{

outSummStat(iris)

}

}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
