% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDendrogram.R
\name{plotDendrogram}
\alias{plotDendrogram}
\title{Plot dendrogram}
\usage{
plotDendrogram(
  dend,
  grp_cnt = NA,
  dend_lbl = NA,
  dend_title = NA,
  dend_xlab = NA,
  dend_ylab = "Distance",
  dist_method = NA,
  aggl_method = NA
)
}
\arguments{
\item{dend}{dend}

\item{grp_cnt}{grp_cnt}

\item{dend_lbl}{dend_lbl}

\item{dend_title}{dend_title}

\item{dend_xlab}{dend_xlab}

\item{dend_ylab}{dend_ylab}

\item{dist_method}{dist_method}

\item{aggl_method}{aggl_method}
}
\description{
Plot dendrogram
}
\details{
...
}
\examples{
\dontrun{
# TBD

}

}
\seealso{
\code{\link{calcQuanClass}}
}
