% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTextFile.R
\name{readTextFile}
\alias{readTextFile}
\title{Read text file into variable}
\usage{
readTextFile(
  file_name,
  file_path = "..",
  file_type = "standard",
  col_names = NA,
  report = FALSE
)
}
\arguments{
\item{file_name}{Name of input file as character array}

\item{file_path}{File path as character array. Setting \code{file_path = ".."}
reads from the current working directory}

\item{file_type}{Indicates what type of file structure is expected. Set
\code{file_type = "WRTDS"} or \code{file_type = "flow"} to use specialized
settings built into the function.  \code{file_type = "standard"} is the default
setting for typical comma or tab delimited text files with the header
as the first line.}

\item{col_names}{Vector of columns names to use for column names in the
returned data set. Set \code{col_names = NA} to use column headers from
input file.}

\item{report}{Summary information about data after reading. Set
\code{report = FALSE} to suppress information.}
}
\value{
data table
}
\description{
This function facilitates reading in text files with standard 
and non-standard headers. The two non-standard file types are 1) USGS' WRTDS
reporting file for estimated concentrations and loads and 2) USGS
flow retrievals.
}
\details{
...
}
\examples{
# TBD

}
\seealso{
\code{\link{calcQuanClass}}
}
