% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupLabeling.R
\name{rnkFlowCorLab}
\alias{rnkFlowCorLab}
\title{compute group labels based on correlation with flow}
\usage{
rnkFlowCorLab(allMat.mn, wq_parm, clus_var, flow_index)
}
\arguments{
\item{allMat.mn}{table of profile means}

\item{wq_parm}{wq_parm}

\item{clus_var}{column names associated with prof_var}

\item{flow_index}{ex_cov_quan}

\item{c.spec}{list that stores specifications for cluster analysis}
}
\value{
table of results
}
\description{
compute group labels based on correlation with flow
}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
