% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupLabeling.R
\name{rnkRowMnLab}
\alias{rnkRowMnLab}
\title{compute group labels based on mean}
\usage{
rnkRowMnLab(c.spec, allMat.mn)
}
\arguments{
\item{c.spec}{list that stores specifications for cluster analysis}

\item{allMat.mn}{table of profile means}
}
\value{
table of results
}
\description{
compute group labels based on mean
}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
