% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupLabeling.R
\name{rnkTrndLab}
\alias{rnkTrndLab}
\title{compute group labels based on comparison of beginning and ending 3-yr time period}
\usage{
rnkTrndLab(allMat.mn, wq_parm, clus_var)
}
\arguments{
\item{allMat.mn}{table of profile means}

\item{wq_parm}{wq_parm}

\item{clus_var}{column names associated with prof_var}
}
\value{
table of results
}
\description{
compute group labels based on comparison of beginning and ending 3-yr time period
}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
