% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSpec.R
\name{setSpec}
\alias{setSpec}
\title{Organize cluster analysis specifications into a list}
\usage{
setSpec(
  projRoot,
  gamRoot,
  wqParm,
  waterBody,
  slayer,
  gamNumbr,
  startYear,
  endYear,
  idVar,
  profVar,
  rowAdj,
  scale,
  grpCnt,
  grpCol,
  statVec,
  statLab,
  monVec,
  monLab,
  filename = NA,
  monOrd = NA,
  yrOrd = NA,
  dataOut = NA,
  subGrps = FALSE,
  addFlow,
  flowFile,
  flowLab = "Flow",
  datSource,
  datFile,
  lineOne,
  grpLab = NA,
  yearType = "calendar"
)
}
\arguments{
\item{projRoot}{root folder for analysis}

\item{gamRoot}{folder where output from baytrends is stored. There should be
subfolders in this folder with names corresponding to the various parameters
of interest. For example, if \code{wqParm} is set to \code{"chla"}, then there should 
be a subdirectory under the \code{gamRoot} folder called \code{chla}.}

\item{wqParm}{parameter to be analyzed}

\item{waterBody}{waterBody}

\item{slayer}{slayer}

\item{gamNumbr}{gamNumbr}

\item{startYear}{First year of data to be used in cluster analysis}

\item{endYear}{Last year of data to be used in cluster analysis}

\item{idVar}{Variable that identifies the rows of the cluster matrix, i.e. items that are clustered}

\item{profVar}{Variable that becomes columns of the cluster matrix, i.e. attribute to cluster}

\item{rowAdj}{rowAdj}

\item{scale}{scale}

\item{grpCnt}{grpCnt}

\item{grpCol}{grpCol}

\item{statVec}{statVec}

\item{statLab}{statLab}

\item{monVec}{monVec}

\item{monLab}{monLab}

\item{filename}{filename=NA}

\item{monOrd}{monOrd=NA,}

\item{yrOrd}{yrOrd=NA,}

\item{dataOut}{dataOut=NA,}

\item{subGrps}{subGrps=FALSE,}

\item{addFlow}{addFlow,}

\item{flowFile}{flowFile,}

\item{flowLab}{flowLab='Flow',}

\item{datSource}{datSource}

\item{datFile}{datFile,}

\item{lineOne}{lineOne,}

\item{grpLab}{grpLab=NA,}

\item{yearType}{yearType = 'calendar'}
}
\value{
list with cluster specifications
}
\description{
Organize user-provided cluster analysis specifications into a
  list. And where specifications are not provided then make some default
  selections.
}
\details{
TBD
}
\examples{
# TBD

}
\seealso{
\code{\link{calcQuanClass}}
}
