% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSpec.R
\name{setSpec}
\alias{setSpec}
\title{Create (or update) specification list for cluster analysis}
\usage{
setSpec(c.spec = list(), ...)
}
\arguments{
\item{c.spec}{list for storing cluster analysis specifications}

\item{...}{specifications needed to run cluster analysis -- see details}
}
\value{
list
}
\description{
Organize user-provided cluster analysis specifications into a
  list. And where specifications are not provided then make some default
  selections.
}
\details{
Below is the list of arguments needed to run a cluster analysis:
\itemize{
\item analysisTitle - Analysis title 
\item analysisDate - Analysis date 
\item datSource - Type of analysis 
\item projFolder - Project folder 
\item gamFolder - GAM folder 
\item datFolder - Input data folder 
\item outFolder - Output data folder 
\item statFile - Station file name 
\item startYear - Start year 
\item endYear - End year 
\item monthGrid - Months included 
\item dayGrid - Days of month 
\item monthAdj - Month adjustment 
\item statVec - Station list 
\item wqParm - Water quality parameter 
\item wqLayer - Layer 
\item gamNumbr - GAM #'s 
\item idVar - Items that are to be clustered
\item profVar - Attribute to cluster by 
\item grpCnt - Number of cluster 
\item dataTrans - Data transformation 
\item dataCenter - Data centering 
\item monthGracePeriod - Grace period 
\item exCovIncl - Include exogenous covariate 
\item exCovFolder - Ex. Cov. Folder 
\item exCovLab - Ex. Cov. Label 
\item exCovFile - Ex. Cov. File 
\item exCovClass - Ex. Cov. # classes 
\item exCovTrans - Ex. Cov. Transformation 
}
}
\examples{
\dontrun{
#TBD

}

}
\seealso{
\code{\link{calcQuanClass}}
}
