% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSpec.R
\name{setSpec}
\alias{setSpec}
\title{Create (or update) specification list for cluster analysis}
\usage{
setSpec(c.spec = list(), ...)
}
\arguments{
\item{c.spec}{list for storing cluster analysis specifications}

\item{...}{specifications needed to run cluster analysis -- see details}
}
\value{
list
}
\description{
Organize user-provided cluster analysis specifications into a
  list. And where specifications are not provided then make some default
  selections.
}
\details{
...
potVar projRoot root folder for analysis
potVar gamRoot folder where output from baytrends is stored. There should be
  subfolders in this folder with names corresponding to the various parameters
  of interest. For example, if \code{wqParm} is set to \code{"chla"}, then there should 
  be a subdirectory under the \code{gamRoot} folder called \code{chla}. 
potVar wqParm parameter to be analyzed
potVar waterBody waterBody
potVar slayer slayer
potVar gamNumbr gamNumbr
potVar startYear First year of data to be used in cluster analysis
potVar endYear Last year of data to be used in cluster analysis
potVar idVar Variable that identifies the rows of the cluster matrix, i.e. items that are clustered 
potVar profVar Variable that becomes columns of the cluster matrix, i.e. attribute to cluster
potVar rowAdj rowAdj
potVar scale  scale
potVar grpCnt grpCnt
potVar grpCol grpCol
potVar statVec statVec
potVar statLab statLab
potVar monVec monVec
potVar monLab monLab
potVar filename filename=NA
potVar monOrd monOrd=NA,
potVar yrOrd yrOrd=NA,
potVar dataOut dataOut=NA,
potVar subGrps subGrps=FALSE,
potVar addFlow addFlow,
potVar flowFile flowFile,
potVar flowLab flowLab='Flow',
potVar datSource datSource
potVar datFile datFile,
potVar lineOne lineOne,
potVar grpLab grpLab=NA,
potVar yearType yearType = 'calendar'
}
\examples{
\dontrun{
#TBD

}

}
\seealso{
\code{\link{calcQuanClass}}
}
