% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUtilities.R
\name{stow}
\alias{stow}
\title{append variables to a list}
\usage{
stow(listvar = list(), v = NA)
}
\arguments{
\item{listvar}{name of target list}

\item{v}{vector of variables to stow into list}
}
\description{
append variables to a list
}
\details{
...
}
\examples{

my_list <- list(a = 4, b = 5, c = "charlie")
pry(my_list, c("a", "c"))

d = "delta"
my_list <- stow(my_list, v = "d")
my_list

}
\seealso{
\code{\link{pry}}
}
\keyword{internal}
