% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblFT1.R
\name{tblFT1}
\alias{tblFT1}
\title{Output formatted table.}
\usage{
tblFT1(
  data,
  tblTitle = NA,
  tblPre_label = NA,
  tblFontName = NA,
  tblFontSize = NA,
  tblTheme = NA
)
}
\arguments{
\item{data}{table data}

\item{tblTitle}{table title}

\item{tblPre_label}{table title pre_label}

\item{tblFontName}{table font}

\item{tblFontSize}{font size}

\item{tblTheme}{table theme}
}
\value{
# may return flextable
}
\description{
This function outputs a table of data with basic formatting.
}
\details{
...
}
\examples{
\dontrun{

tblFT1(head(iris))

}

}
\seealso{
\code{\link{calcQuanClass}}
}
