% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblFT1.R
\name{tblFT1}
\alias{tblFT1}
\title{Output formatted table.}
\usage{
tblFT1(
  data,
  tbl_title = NA,
  tbl_pre_label = NA,
  tbl_font_name = NA,
  tbl_font_size = NA,
  tbl_theme = NA
)
}
\arguments{
\item{data}{table data}

\item{tbl_title}{table title}

\item{tbl_pre_label}{table title pre_label}

\item{tbl_font_name}{table font}

\item{tbl_font_size}{font size}

\item{tbl_theme}{table theme}
}
\value{
# may return flextable
}
\description{
This function outputs a table of data with basic formatting.
}
\details{
...
}
\examples{
\dontrun{

tblFT1(head(iris))

}

}
\seealso{
\code{\link{calcQuanClass}}
}
