% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupLabeling.R
\name{yeargrp_lab}
\alias{yeargrp_lab}
\title{compute group labels based on years within group}
\usage{
yeargrp_lab(allMat.grp, wq_parm, clus_var)
}
\arguments{
\item{wq_parm}{wq_parm}

\item{clus_var}{column names associated with prof_var}

\item{allMat.mn}{table of profile means}
}
\value{
table of results
}
\description{
compute group labels based on years within group
}
\seealso{
\code{\link{calcQuanClass}}
}
\keyword{internal}
